CA-781 Moved participantInfromation type
