+++ src/types/participantInformation.ts
@@ -1,0 +1,7 @@
+export interface ParticipantInformation {
+  caseReference: string;
+  displayName: string;
+  partyType: string;
+  userId: number | null;
+  videoConferenceRoomName: string;
+}
